fix long number

+++ is_prime_prog/main.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 
-bool is_prime(unsigned x)
+bool is_prime(unsigned long x)
 {
     if (x <= 1 || x%2 == 0 && x != 2) {
         return false;
     }
 
-    for (int i = 3; i <= x/2; i+=2) {
+    for (unsigned i = 3; i <= x/2; i+=2) {
         if (x % i == 0) {
             return false;
         }
@@ -16,7 +16,7 @@ bool is_prime(unsigned x)
 
 int main(int argc, char *argv[])
 {
-    unsigned n;
+    unsigned long n;
     while(std::cin >> n)
     {
         if(is_prime(n))
